refactor: Move source code setup fixtures to conftest
Also make them return a Dict instead of RuntimeComposeFileModel. 
s
